update 2.48 removeComments

+++ com.gongpingjia.carplay/com.gongpingjia.carplay.service/src/main/java/com/gongpingjia/carplay/controller/MessageController.java
@@ -35,8 +35,11 @@ public class MessageController {
 	 * 
 	 * @param userId
 	 *            访问者的userId
-	 * @param request
-	 *            请求参数
+	 * @param token
+	 *            访问者的token
+	 * @param ignore
+	 * 
+	 * @param limit
 	 * 
 	 * @return 活动申请列表信息
 	 * 
@@ -88,8 +91,15 @@ public class MessageController {
 	 * 
 	 * @param userId
 	 *            用户Id
-	 * @param request
-	 *            请求参数
+	 * @param token
+	 * 
+	 * @param type
+	 *            消息类型，comment与application
+	 * 
+	 * @param ignore
+	 * 
+	 * @param limit
+	 * 
 	 * @return 消息列表
 	 */
 	@RequestMapping(value = "/user/{userId}/message/list", method = RequestMethod.GET)
@@ -99,7 +109,12 @@ public class MessageController {
 		LOG.debug("==> getMessageList");
 
 		try {
+			if (!type.equals("comment") && !type.equals("application")) {
+				LOG.warn("invalid params");
+				throw new ApiException("消息类型有误");
+			}
 			checker.checkUserInfo(userId, token);
+
 			return messageService.getMessageList(userId, type, ignore, limit);
 		} catch (ApiException e) {
 
@@ -111,8 +126,15 @@ public class MessageController {
 	/**
 	 * 2.44 提交反馈信息
 	 * 
-	 * @param request
-	 *            请求参数
+	 * @param userId
+	 * 
+	 * @param token
+	 * 
+	 * @param content
+	 *            反馈信息
+	 * 
+	 * @param photos
+	 *            反馈图片ID
 	 * 
 	 * @return 提交成功
 	 */
@@ -123,6 +145,10 @@ public class MessageController {
 		LOG.debug("==> submitFeedback");
 
 		try {
+			if (content == null || (photos != null && photos.length > 3)) {
+				LOG.warn("invalid params");
+				throw new ApiException("请输入需要反馈的信息");
+			}
 			checker.checkUserInfo(userId, token);
 			return messageService.submitFeedback(userId, content, photos);
 		} catch (ApiException e) {
@@ -134,8 +160,12 @@ public class MessageController {
 	/**
 	 * 2.47批量删除消息
 	 * 
-	 * @param request
-	 *            请求参数
+	 * @param userId
+	 * 
+	 * @param token
+	 * 
+	 * @param messages
+	 *            消息Id
 	 *
 	 * @return 删除成功
 	 */
@@ -146,11 +176,47 @@ public class MessageController {
 
 		try {
 			checker.checkUserInfo(userId, token);
+			if (messages == null || messages.length == 0) {
+				LOG.warn("invalid params");
+				throw new ApiException("输入参数有误");
+			}
 			return messageService.removeMessages(userId, messages);
 		} catch (ApiException e) {
 			LOG.warn(e.getMessage(), e);
 			return ResponseDo.buildFailureResponse(e.getMessage());
 		}
+	}
+
+	/**
+	 * 2.48 批量删除评论
+	 * 
+	 * @param userId
+	 * 
+	 * @param token
+	 * 
+	 * @param comments
+	 *            评论ID
+	 * 
+	 * @return 删除成功
+	 * 
+	 */
+	@RequestMapping(value = "/comment/remove", method = RequestMethod.POST)
+	public ResponseDo removeComments(@RequestParam("userId") String userId, @RequestParam("token") String token,
+			@RequestParam("comments") String[] comments) {
+		LOG.debug("==>removeComments");
+
+		try {
+			checker.checkUserInfo(userId, token);
+			if (comments == null || comments.length == 0) {
+				LOG.warn("invalid params");
+				throw new ApiException("输入参数有误");
+			}
+			return messageService.removeComments(userId, comments);
+		} catch (ApiException e) {
+			LOG.warn(e.getMessage(), e);
+			return ResponseDo.buildFailureResponse(e.getMessage());
+		}
+
 	}
 
 }
